Move validation to shared-utilities
